# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/form-responses": {
      "get": {
        "operationId": "Forms_listResponses",
        "parameters": [
          {
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "formKey",
            "required": false,
            "schema": {
              "pattern": "^[a-z][a-z0-9_]{0,60}$",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "draft",
                "submitted"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subjectPersonId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      },
      "post": {
        "operationId": "Forms_createResponse",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "contextId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "contextType": {
                    "maxLength": 40,
                    "type": "string"
                  },
                  "dueDate": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "formDefinitionId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "formKey": {
                    "pattern": "^[a-z][a-z0-9_]{0,60}$",
                    "type": "string"
                  },
                  "notify": {
                    "type": "boolean"
                  },
                  "respondentPersonId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "subjectPersonId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/form-responses/{id}": {
      "get": {
        "operationId": "Forms_getResponse",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/form-responses/{id}/draft": {
      "put": {
        "operationId": "Forms_draft",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "answers": {
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        {
                          "enum": [
                            "null"
                          ],
                          "nullable": true
                        }
                      ]
                    },
                    "type": "object"
                  }
                },
                "required": [
                  "answers"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/form-responses/{id}/submit": {
      "post": {
        "operationId": "Forms_submit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "answers": {
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        {
                          "enum": [
                            "null"
                          ],
                          "nullable": true
                        }
                      ]
                    },
                    "type": "object"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/form-responses/{id}/validate": {
      "post": {
        "operationId": "Forms_validate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "answers": {
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        {
                          "enum": [
                            "null"
                          ],
                          "nullable": true
                        }
                      ]
                    },
                    "type": "object"
                  }
                },
                "required": [
                  "answers"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/forms": {
      "get": {
        "operationId": "Forms_list",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "draft",
                "published",
                "archived"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "latest",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      },
      "post": {
        "operationId": "Forms_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "key": {
                    "pattern": "^[a-z][a-z0-9_]{0,60}$",
                    "type": "string"
                  },
                  "kind": {
                    "default": "generic",
                    "enum": [
                      "generic",
                      "review",
                      "survey",
                      "onboarding",
                      "request",
                      "feedback360"
                    ],
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "schema": {}
                },
                "required": [
                  "key",
                  "name",
                  "schema"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea una definizione di form (bozza) a partire da uno schema dichiarativo",
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/forms/{id}": {
      "get": {
        "operationId": "Forms_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      },
      "patch": {
        "operationId": "Forms_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  },
                  "schema": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/forms/{id}/publish": {
      "post": {
        "operationId": "Forms_publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
    "/api/v1/forms/{id}/versions": {
      "post": {
        "operationId": "Forms_newVersion",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "forms"
        ]
      }
    },
...
  },
  "info": {
    "title": "WorkingBetter API",
    "description": "API REST unica per web, mobile e connettori (ADR-0005). Errori in formato RFC 9457 (application/problem+json). I body e le query sono descritti dagli stessi schemi Zod usati per la validazione (ADR-0009).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "URI del tipo di errore (about:blank per gli errori generici)"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "description": "Identificativo della richiesta (x-request-id)"
          },
          "code": {
            "type": "string",
            "description": "Codice applicativo stabile (ErrorCodes in @wb/shared)"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "path",
                "message"
              ],
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}